[IMP] 'custom_account': Reorganización código rappel y añadida vista para mostrar en ficha de cliente con la configuración del periodo del rappel y la cantidad que lleva conseguido
