+++ project-addons/custom_partner/__init__.py
@@ -21,3 +21,4 @@
 
 from . import partner
 from . import stock
+from . import rappel
