+++ project-addons/custom_partner/rappel.py
@@ -1,0 +1,453 @@
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+#    Copyright (C) 2015 Comunitea Servicios Tecnológicos All Rights Reserved
+#    $Omar Castiñeira Saavedra <omar@pcomunitea.com>$
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as published
+#    by the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp import models, fields, api, exceptions, osv, tools, _
+from datetime import datetime
+from calendar import monthrange
+from dateutil.relativedelta import relativedelta
+
+
+class rappel_calculated(models.Model):
+
+    _inherit = 'rappel.calculated'
+
+    goal_percentage = fields.Float("Goal Percentage")
+
+
+class ResPartnerRappelRel(models.Model):
+
+    _inherit = "res.partner.rappel.rel"
+
+    @api.multi
+    def _get_next_period(self):
+        period = super(ResPartnerRappelRel, self)._get_next_period()
+        if period and str(period[0]) == self.last_settlement_date:
+            period[0] += relativedelta(days=1)
+            period[1] += relativedelta(days=1)
+        return period
+
+    @api.multi
+    def _get_invoices(self, period, products):
+        res = super(ResPartnerRappelRel, self)._get_invoices(period, products)
+
+        self.ensure_one()
+        invoices = self.env['account.invoice'].search(
+            [('type', '=', 'out_invoice'),
+             ('date_invoice', '>=', period[0]),
+             ('date_invoice', '<=', period[1]),
+             ('state', 'in', ['open', 'paid']),
+             ('commercial_partner_id', '=', self.partner_id.id)])
+        refunds = self.env['account.invoice'].search(
+            [('type', '=', 'out_refund'),
+             ('date_invoice', '>=', period[0]),
+             ('date_invoice', '<=', period[1]),
+             ('state', 'in', ['open', 'paid']),
+             ('commercial_partner_id', '=', self.partner_id.id)])
+
+        # Si el rappel afecta al catalago entero, no hacer la comprobacion por producto
+        if self.rappel_id.global_application:
+            refund_lines = self.env['account.invoice.line'].search(
+                [('invoice_id', 'in', [x.id for x in refunds]),
+                 ('no_rappel', '=', False)])
+            invoice_lines = self.env['account.invoice.line'].search(
+                [('invoice_id', 'in', [x.id for x in invoices]),
+                 ('no_rappel', '=', False)])
+        else:
+            refund_lines = self.env['account.invoice.line'].search(
+                [('invoice_id', 'in', [x.id for x in refunds]),
+                 ('product_id', 'in', products),
+                 ('no_rappel', '=', False)])
+            invoice_lines = self.env['account.invoice.line'].search(
+                [('invoice_id', 'in', [x.id for x in invoices]),
+                 ('product_id', 'in', products),
+                 ('no_rappel', '=', False)])
+
+        return invoice_lines, refund_lines
+
+    @api.multi
+    def _calculate_qty_picking(self):
+        picking_obj = self.env['stock.picking']
+        move_obj = self.env['stock.move']
+        products = self.rappel_id.get_products()
+        period = self._get_next_period()
+        if period:
+            picking_ids = picking_obj.search([('date_done', '>=', period[0].strftime("%Y-%m-%d")),
+                                              ('date_done', '<=', period[1].strftime("%Y-%m-%d")),
+                                              ('state', '=', 'done'),
+                                              ('invoice_state', '=', '2binvoiced'),
+                                              ('partner_id', 'child_of', [self.partner_id.id])])
+
+            picking_lines = move_obj.search([('picking_id', 'in', picking_ids.ids),
+                                             ('product_id', 'in', products)])
+
+            price_subtotal_lines = picking_lines.mapped('procurement_id.sale_line_id.price_subtotal')
+            amount_total = sum([x for x in price_subtotal_lines])
+            return amount_total
+
+    @api.model
+    def compute(self, period, invoice_lines, refund_lines, tmp_model=False):
+        goal_percentage = 0
+        for rappel in self:
+            rappel_info = {'rappel_id': rappel.rappel_id.id,
+                           'partner_id': rappel.partner_id.id,
+                           'date_start': period[0],
+                           'amount': 0.0,
+                           'date_end': period[1]}
+            total_rappel = 0.0
+            if rappel.rappel_id.calc_mode == 'fixed':
+                if rappel.rappel_id.calc_amount == 'qty':
+                    total_rappel = rappel.rappel_id.fix_qty
+                else:
+                    total = sum([x.price_subtotal for x in invoice_lines]) - \
+                        sum([x.price_subtotal for x in refund_lines])
+                    if total:
+                        total_rappel = total * rappel.rappel_id.fix_qty / 100.0
+                    rappel_info["curr_qty"] = total
+                    rappel_info["curr_qty_pickings"] = rappel._calculate_qty_picking()
+
+                rappel_info['amount'] = total_rappel
+            else:
+                field = ''
+                if rappel.rappel_id.qty_type == 'value':
+                    field = 'price_subtotal'
+                else:
+                    field = 'quantity'
+                total = sum([x[field] for x in invoice_lines]) - \
+                    sum([x[field] for x in refund_lines])
+                rappel_info["curr_qty"] = total
+                rappel_info["curr_qty_pickings"] = rappel._calculate_qty_picking()
+                if total:
+                    section = self.env['rappel.section'].search(
+                        [('rappel_id', '=', rappel.rappel_id.id),
+                         ('rappel_from', '<=', total),
+                         ('rappel_until', '>=', total)])
+                    if not section:
+                        section = self.env['rappel.section'].search(
+                            [('rappel_id', '=', rappel.rappel_id.id),
+                             ('rappel_from', '<=', total),
+                             ('rappel_until', '=', False)],
+                            order='rappel_from desc', limit=1)
+                    if section:
+                        goal_percentage = 100
+                    else:
+                        # Check if goal percentage is more than 80% to get the rappel
+                        section = self.env['rappel.section'].search(
+                            [('rappel_id', '=', rappel.rappel_id.id),
+                             ('rappel_from', '<=', total/0.8),
+                             ('rappel_from', '>', total)])
+                        if section.rappel_from:
+                            goal_percentage = (total / section.rappel_from) * 100
+                        else:
+                            goal_percentage = 0
+
+                    if not section:
+                        rappel_info['amount'] = 0.0
+                    else:
+                        rappel_info['section_id'] = section.id
+                        section = section[0]
+                        if rappel.rappel_id.calc_amount == 'qty':
+                            total_rappel = section.percent
+                        else:
+                            total_rappel = total * \
+                                section.percent / 100.0
+                            rappel_info['amount'] = total_rappel
+                else:
+                    rappel_info['amount'] = 0.0
+
+            if period[1] <= fields.Date.from_string(fields.Date.today()):
+                if total_rappel:
+                    self.env['rappel.calculated'].create({
+                        'partner_id': rappel.partner_id.id,
+                        'date_start': period[0],
+                        'date_end': period[1],
+                        'quantity': total_rappel,
+                        'rappel_id': rappel.rappel_id.id,
+                        'goal_percentage': goal_percentage
+                    })
+                rappel.last_settlement_date = period[1]
+            else:
+                if tmp_model and rappel_info:
+                    self.env['rappel.current.info'].create(rappel_info)
+
+        return True
+
+
+class rappel(models.Model):
+
+    _inherit = 'rappel'
+    brand_ids = fields.Many2many('product.brand', 'rappel_product_brand_rel',
+                                 'rappel_id', 'product_brand_id', 'Brand')
+    discount_voucher = fields.Boolean('Discount voucher')
+
+    @api.multi
+    def get_products(self):
+        product_obj = self.env['product.product']
+        product_ids = self.env['product.product']
+        for rappel in self:
+            if not rappel.global_application:
+                if rappel.product_id:
+                    product_ids += rappel.product_id
+                elif rappel.brand_ids:
+                    product_ids += product_obj.search(
+                        [('product_brand_id', 'in', rappel.brand_ids.ids)])
+                elif rappel.product_categ_id:
+                    product_ids += product_obj.search(
+                        [('categ_id', '=', rappel.product_categ_id.id)])
+            else:
+                product_ids += product_obj.search([])
+        return product_ids.ids
+
+    @api.constrains('global_application', 'product_id', 'brand_ids', 'product_categ_id')
+    def _check_application(self):
+        if not self.global_application and not self.product_id \
+                and not self.product_categ_id and not self.brand_ids:
+            raise exceptions. \
+                ValidationError(_('Product, brand and category are empty'))
+
+    @api.model
+    def update_partner_rappel_pricelist(self):
+        partner_obj = self.env['res.partner']
+        pricelist_obj = self.env['product.pricelist']
+        rappel_obj = self.env['rappel']
+        partner_rappel_obj = self.env['res.partner.rappel.rel']
+        now = datetime.now()
+        now_str = now.strftime("%Y-%m-%d")
+
+        pricelist_1 = tuple(pricelist_obj.search([('name', 'ilike', 'PVP%52,5')]).ids)
+        pricelist_2 = tuple(pricelist_obj.search([('name', 'ilike', 'PVP%55')]).ids)
+        rappel_pricelist_1 = rappel_obj.search([('name', 'ilike', 'Vale ahorro%5')]).id
+        rappel_pricelist_2 = rappel_obj.search([('name', 'ilike', 'Vale ahorro%10')]).id
+
+        partner_pricelist_1 = tuple(partner_obj.search([('property_product_pricelist', 'in', pricelist_1),
+                                                        ('prospective', '=', False), ('active', '=', True),
+                                                        ('is_company', '=', True), ('parent_id', '=', False)]).ids)
+        partner_pricelist_2 = tuple(partner_obj.search([('property_product_pricelist', 'in', pricelist_2),
+                                                        ('prospective', '=', False), ('active', '=', True),
+                                                        ('is_company', '=', True), ('parent_id', '=', False)]).ids)
+        partner_rappel_1 = tuple(partner_rappel_obj.search([('rappel_id', '=', rappel_pricelist_1),
+                                                            '|',  ('date_end', '=', False),
+                                                            ('date_end', '>', now), ('date_start', '<=', now_str)]).mapped('partner_id.id'))
+        partner_rappel_2 = tuple(partner_rappel_obj.search([('rappel_id', '=', rappel_pricelist_2),
+                                                            '|', ('date_end', '=', False),
+                                                            ('date_end', '>', now), ('date_start', '<=', now_str)]).mapped('partner_id.id'))
+
+        end_actual_month = now.strftime("%Y-%m") + '-' + str(monthrange(now.year, now.month)[1])
+        start_next_month = (now + relativedelta(months=1)).strftime("%Y-%m") + '-01'
+
+        # ---------------------
+        # TARIFA 1 : PVP 52,5%
+        # ---------------------
+
+        #  Clientes que faltan en el rappel -> Se crean dos entradas en el rappel:
+        #      - Una para liquidar en el mes actual
+        #      - Otra que empiece en fecha 1 del mes siguiente
+        add_partners = set(partner_pricelist_1) - set(partner_rappel_1)
+        if add_partners:
+            new_line1 = {'rappel_id': rappel_pricelist_1, 'periodicity': 'monthly',
+                         'date_start': now_str, 'date_end': end_actual_month}
+            new_line2 = {'rappel_id': rappel_pricelist_1, 'periodicity': 'monthly',  'date_start': start_next_month}
+            for partner in add_partners:
+                new_line1.update({'partner_id': partner})
+                partner_rappel_obj.create(new_line1)
+                new_line2.update({'partner_id': partner})
+                partner_rappel_obj.create(new_line2)
+
+        # Clientes a los que ya no les corresponde el rappel -> Se actualiza fecha fin con la fecha actual
+        remove_partners = set(partner_rappel_1) - set(partner_pricelist_1)
+        if remove_partners:
+            vals = {'date_end': now_str}
+            partner_to_update = partner_rappel_obj.search([('rappel_id', '=', rappel_pricelist_1),
+                                                           ('partner_id', 'in', tuple(remove_partners)),
+                                                           '|',  ('date_end', '=', False),
+                                                           ('date_end', '>', now), ('date_start', '<=', now_str)])
+            partner_to_update.write(vals)
+
+        # ---------------------
+        # TARIFA 2 : PVP 55%
+        # ---------------------
+
+        #  Clientes que faltan en el rappel -> Se crean dos entradas en el rappel:
+        #      - Una para liquidar en el mes actual
+        #      - Otra que empiece en fecha 1 del mes siguiente
+        add_partners = set(partner_pricelist_2) - set(partner_rappel_2)
+        if add_partners:
+            new_line1 = {'rappel_id': rappel_pricelist_2, 'periodicity': 'monthly',
+                         'date_start': now_str, 'date_end': end_actual_month}
+            new_line2 = {'rappel_id': rappel_pricelist_2, 'periodicity': 'monthly', 'date_start': start_next_month}
+            for partner in add_partners:
+                new_line1.update({'partner_id': partner})
+                partner_rappel_obj.create(new_line1)
+                new_line2.update({'partner_id': partner})
+                partner_rappel_obj.create(new_line2)
+
+        # Clientes a los que ya no les corresponde el rappel -> Se actualiza fecha fin con la fecha actual
+        remove_partners = set(partner_rappel_2) - set(partner_pricelist_2)
+        if remove_partners:
+            update_date = {'date_end': now_str}
+            partner_to_update = partner_rappel_obj.search([('rappel_id', '=', rappel_pricelist_2),
+                                                           ('partner_id', 'in', tuple(remove_partners)),
+                                                           '|', ('date_end', '=', False),
+                                                           ('date_end', '>', now), ('date_start', '<=', now_str)])
+            partner_to_update.write(update_date)
+
+
+class RappelInvoice(models.TransientModel):
+
+    _inherit = "rappel.invoice.wzd"
+
+    @api.multi
+    def action_invoice(self):
+        res = super(RappelInvoice, self).action_invoice()
+        compute_rappel_obj = self.env["rappel.calculated"]
+        for rappel in compute_rappel_obj.browse(self.env.context["active_ids"]):
+            if rappel.quantity <= 0:
+                continue
+            if rappel.invoice_id:
+                invoice = rappel.invoice_id
+                if not invoice.payment_mode_id \
+                        or not invoice.partner_bank_id \
+                        or not invoice.section_id:
+                    rappel.invoice_id.write({'payment_mode_id': rappel.partner_id.customer_payment_mode.id,
+                                             'partner_bank_id': rappel.partner_id.bank_ids and
+                                                                    rappel.partner_id.bank_ids[0].id or False,
+                                             'section_id': rappel.partner_id.section_id.id})
+        return res
+
+
+class RappelCurrentInfo(models.Model):
+
+    _inherit = "rappel.current.info"
+
+    curr_qty_pickings = fields.Float("Qty pending invoice", readonly=True,
+                                     help="Qty estimation in pickings pending to be invoiced (shipping cost and"
+                                          "product with no-rappel in the order are not verified)")
+
+    @api.model
+    def send_rappel_info_mail(self):
+        mail_pool = self.env['mail.mail']
+        mail_ids = self.env['mail.mail']
+        partner_pool = self.env['res.partner'].search([('rappel_ids', '!=', '')])
+        for partner in partner_pool:
+            partner_list = []
+            partner_list.append(partner.id)
+            pool_partners = self.search([('partner_id', '=', partner.id)])
+            send = False
+            if pool_partners:
+
+                values = {}
+                for rappel in pool_partners:
+
+                    date_end = datetime.strptime(str(rappel.date_end), '%Y-%m-%d')
+                    date_start = datetime.strptime(str(rappel.date_start), '%Y-%m-%d')
+                    today = datetime.strptime(str(fields.Date.today()), '%Y-%m-%d')
+
+                    for rappel_timing in rappel.rappel_id.advice_timing_ids:
+
+                        if rappel_timing.advice_timing == 'fixed':
+                            timing = (date_end - today).days
+                            if timing == rappel_timing.timing:
+                                send = True
+
+                        if rappel_timing.advice_timing == 'variable':
+
+                            timing = (date_end - date_start).days*rappel_timing.timing/100
+                            timing2= (today - date_start).days
+
+                            if timing == timing2:
+                                send = True
+
+                        if send == True and rappel.curr_qty:
+                            if values.get(partner.id):
+                                values[partner.id].append ({
+                                    'concepto': rappel.rappel_id.name,
+                                    'date_start': date_start.strftime('%d/%m/%Y'),
+                                    'date_end': date_end.strftime('%d/%m/%Y'),
+                                    'advice_timing': rappel_timing.advice_timing,
+                                    'timing': rappel_timing.timing,
+                                    'curr_qty': rappel.curr_qty,
+                                    'section_goal': rappel.section_goal,
+                                    'section_id': rappel.section_id,
+                                    'amount': rappel.amount
+                                })
+                            else:
+                                values[partner.id] = [{
+                                    'concepto': rappel.rappel_id.name,
+                                    'date_start': date_start.strftime('%d/%m/%Y'),
+                                    'date_end': date_end.strftime('%d/%m/%Y'),
+                                    'advice_timing': rappel_timing.advice_timing,
+                                    'timing': rappel_timing.timing,
+                                    'curr_qty': rappel.curr_qty,
+                                    'section_goal': rappel.section_goal,
+                                    'section_id': rappel.section_id,
+                                    'amount': rappel.amount
+                                }]
+                        send = False
+
+                if values.get(partner.id):
+                    template = self.env.ref('rappel.rappel_mail_advice')
+                    ctx = dict(self._context)
+                    ctx.update({
+                        'partner_email': partner.email,
+                        'partner_id': partner.id,
+                        'partner_lang': partner.lang,
+                        'partner_name': partner.name,
+                        'mail_from': self.env.user.company_id.email,
+                        'values': values[partner.id]
+                    })
+
+                    mail_id = template.with_context(ctx).send_mail(rappel.partner_id.id)
+                    mail_ids += mail_pool.browse(mail_id)
+        if mail_ids:
+            mail_ids.send()
+
+
+class PartnerRappelInfo(models.Model):
+
+    _name = "partner.rappel.info"
+    _auto = False
+    _order = 'partner_id, rappel_id desc'
+    # _table = 'partner_rappel_info'
+
+    partner_id = fields.Many2one('res.partner', 'Partner')
+    rappel_id = fields.Many2one('rappel', 'Rappel')
+    start_rappel = fields.Date('Start rappel')
+    end_rappel = fields.Date('End rappel')
+    last_settlement_date = fields.Date('Last settlement date')
+    start_current_info = fields.Date('Start current period')
+    end_current_info = fields.Date('End current period')
+    current_amount = fields.Float('Current rappel amount')
+    discount_voucher = fields.Boolean('Discount voucher')
+
+    def init(self, cr):
+        # self._table = partner_rappel_info
+        tools.drop_view_if_exists(cr, self._table)
+        cr.execute("""
+            CREATE OR REPLACE VIEW %s AS (
+                SELECT  rprr.id, rci.partner_id, rci.rappel_id, r.discount_voucher,
+                        rprr.date_start as start_rappel, rprr.date_end as end_rappel, last_settlement_date,
+                        rci.date_start as start_current_info, rci.date_end as end_current_info, 
+                        rci.amount as current_amount
+                FROM rappel_current_info rci
+                JOIN res_partner_rappel_rel rprr ON rprr.partner_id = rci.partner_id and rprr.rappel_id = rci.rappel_id
+                JOIN rappel r ON r.id = rci.rappel_id 
+            )""" % self._table)
+
+
